Added tags

+++ src/notes/providers/notes.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Note } from '../note.entity';
-import { FindOptionsOrderValue, Repository } from 'typeorm';
+import { FindOptionsOrderValue, In, Repository } from 'typeorm';
 import { CreateNoteDto } from '../dto/create-note.dto';
 import { UpdateNoteDto } from '../dto/update-note.dto';
 import { AuthService } from 'src/users/providers/auth.service';
@@ -37,16 +37,33 @@ export class NotesService {
     createNoteDto: CreateNoteDto,
   ): Promise<void> {
     const user = await this.authService.getUserFromAuthHeader(authHeader);
+    let note;
 
     if (!user) {
       throw new NotFoundException('User not found');
     }
 
-    const note = this.noteRepository.create({
-      title: createNoteDto.title,
-      content: createNoteDto.content,
-      user,
-    });
+    if (createNoteDto.tagsIds && createNoteDto.tagsIds.length > 0) {
+      const tags = await this.noteRepository.find({
+        where: { id: In(createNoteDto.tagsIds) },
+      });
+
+      if (tags.length !== createNoteDto.tagsIds.length) {
+        throw new NotFoundException('One or more tags not found');
+      }
+      note = await this.noteRepository.create({
+        title: createNoteDto.title,
+        content: createNoteDto.content,
+        tags,
+        user,
+      });
+    } else {
+      note = this.noteRepository.create({
+        title: createNoteDto.title,
+        content: createNoteDto.content,
+        user,
+      });
+    }
 
     await this.noteRepository.save(note);
   }
@@ -56,15 +73,60 @@ export class NotesService {
     id: number,
     updateNoteDto: UpdateNoteDto,
   ): Promise<void> {
-    const { title, content, isActive } = updateNoteDto;
-    await this.noteRepository.update(id, {
-      title,
-      content,
-      isActive,
+    const { title, content, isActive, tagsIds } = updateNoteDto;
+
+    const user = await this.authService.getUserFromAuthHeader(authHeader);
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    const notes = await this.noteRepository.find({
+      where: { user: { id: user.id } },
     });
+
+    if (!notes.some((note) => note.id === id)) {
+      throw new NotFoundException('Note does not belong to provided user');
+    }
+
+    if (tagsIds && tagsIds.length > 0) {
+      const tags = await this.noteRepository.find({
+        where: { id: In(tagsIds) },
+      });
+
+      if (tags.length !== tagsIds.length) {
+        throw new NotFoundException('One or more tags not found');
+      }
+      await this.noteRepository.update(id, {
+        title,
+        content,
+        isActive,
+        tags,
+      });
+    } else {
+      await this.noteRepository.update(id, {
+        title,
+        content,
+        isActive,
+      });
+    }
   }
 
   async remove(authHeader: string, id: number): Promise<void> {
+    const user = await this.authService.getUserFromAuthHeader(authHeader);
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    const notes = await this.noteRepository.find({
+      where: { user: { id: user.id } },
+    });
+
+    if (!notes.some((note) => note.id === id)) {
+      throw new NotFoundException('Note does not belong to provided user');
+    }
+
     await this.noteRepository.delete(id);
   }
 }
